Refactoring: server.js
