+++ server/service/index.js
@@ -5,13 +5,17 @@ const path = require('path');
 const fs = require('fs');
 const exec = require('child_process').exec;
 const request = require('request');
+
+
+
+
 /**
  * Save a PDF file, and return the saved path.
  */
 module.exports.savePDF = (fileName, content) => {
     return new Promise((resolve, reject) => {
 
-        const dataPath = path.resolve(__dirname, 'server-data');
+        const dataPath = path.resolve(__dirname, '..', 'server-data');
         if (!fs.existsSync(dataPath)) {
                 fs.mkdirSync(dataPath);
         }
@@ -36,9 +40,7 @@ module.exports.analyzePDF = (pdfPath) => {
 
     }).then(() => {
 
-        // Prepare pdfreader.jar
-
-        const exists = fs.existsSync(path.resolve(__dirname, 'pdfextract.jar'));
+        const exists = fs.existsSync(path.resolve(__dirname, '..', 'pdfextract.jar'));
         if (exists) {
             return;
         }
@@ -57,7 +59,7 @@ module.exports.analyzePDF = (pdfPath) => {
                     reject(err);
                 }
 
-                fs.writeFileSync(path.resolve(__dirname, 'pdfextract.jar'), buf);
+                fs.writeFileSync(path.resolve(__dirname, '..', 'pdfextract.jar'), buf);
 
                 resolve();
             });
@@ -65,13 +67,11 @@ module.exports.analyzePDF = (pdfPath) => {
 
     }).then(() => {
 
-        const jarPath = path.resolve(__dirname, 'pdfextract.jar');
-        // java -classpath pdfextract.jar PDFExtractor xxx.pdf -text -bounding
+        const jarPath = path.resolve(__dirname, '..', 'pdfextract.jar');
         const cmd = `java -classpath ${jarPath} PDFExtractor ${pdfPath} -text -bounding`;
         return execCommand(cmd);
 
     }).then(({ stdout, stderr }) => {
-
         return stdout;
     });
 }
